add logger function, rename folder
